+++ lib/src/clients/serviceinfo/models/service.rs
