+++ lib/src/clients/tes/models/service_type.rs
@@ -10,42 +10,30 @@
 
 #![allow(unused_imports)]
 #![allow(clippy::empty_docs)]
-use crate::tes::models;
+use crate::clients::tes::models;
 use serde::{Deserialize, Serialize};
 
-/// TesTaskLog : TaskLog describes logging information related to a Task.
+/// ServiceType : Type of a GA4GH service
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
-pub struct TesTaskLog {
-    /// Logs for each executor
-    #[serde(rename = "logs")]
-    pub logs: Vec<models::TesExecutorLog>,
-    /// Arbitrary logging metadata included by the implementation.
-    #[serde(rename = "metadata", skip_serializing_if = "Option::is_none")]
-    pub metadata: Option<std::collections::HashMap<String, String>>,
-    /// When the task started, in RFC 3339 format.
-    #[serde(rename = "start_time", skip_serializing_if = "Option::is_none")]
-    pub start_time: Option<String>,
-    /// When the task ended, in RFC 3339 format.
-    #[serde(rename = "end_time", skip_serializing_if = "Option::is_none")]
-    pub end_time: Option<String>,
-    /// Information about all output files. Directory outputs are flattened into separate items.
-    #[serde(rename = "outputs")]
-    pub outputs: Vec<models::TesOutputFileLog>,
-    /// System logs are any logs the system decides are relevant, which are not tied directly to an Executor process. Content is implementation specific: format, size, etc.  System logs may be collected here to provide convenient access.  For example, the system may include the name of the host where the task is executing, an error message that caused a SYSTEM_ERROR state (e.g. disk is full), etc.  System logs are only included in the FULL task view.
-    #[serde(rename = "system_logs", skip_serializing_if = "Option::is_none")]
-    pub system_logs: Option<Vec<String>>,
+pub struct ServiceType {
+    /// Namespace in reverse domain name format. Use `org.ga4gh` for implementations compliant with official GA4GH specifications. For services with custom APIs not standardized by GA4GH, or implementations diverging from official GA4GH specifications, use a different namespace (e.g. your organization's reverse domain name).
+    #[serde(rename = "group")]
+    pub group: String,
+    /// Name of the API or GA4GH specification implemented. Official GA4GH types should be assigned as part of standards approval process. Custom artifacts are supported.
+    #[serde(rename = "artifact")]
+    pub artifact: String,
+    /// Version of the API or specification. GA4GH specifications use semantic versioning.
+    #[serde(rename = "version")]
+    pub version: String,
 }
 
-impl TesTaskLog {
-    /// TaskLog describes logging information related to a Task.
-    pub fn new(logs: Vec<models::TesExecutorLog>, outputs: Vec<models::TesOutputFileLog>) -> TesTaskLog {
-        TesTaskLog {
-            logs,
-            metadata: None,
-            start_time: None,
-            end_time: None,
-            outputs,
-            system_logs: None,
+impl ServiceType {
+    /// Type of a GA4GH service
+    pub fn new(group: String, artifact: String, version: String) -> ServiceType {
+        ServiceType {
+            group,
+            artifact,
+            version,
         }
     }
 }
