+++ lib/src/clients/tes/models/tes_resources.rs
@@ -10,7 +10,7 @@
 
 #![allow(unused_imports)]
 #![allow(clippy::empty_docs)]
-use crate::tes::models;
+use crate::clients::tes::models;
 use serde::{Deserialize, Serialize};
 
 /// TesResources : Resources describes the resources requested by a task.
