+++ lib/src/clients/tes/models/tes_service_type.rs
@@ -10,42 +10,40 @@
 
 #![allow(unused_imports)]
 #![allow(clippy::empty_docs)]
-use crate::tes::models;
+use crate::clients::tes::models;
 use serde::{Deserialize, Serialize};
 
-/// TesOutput : Output describes Task output files.
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
-pub struct TesOutput {
-    /// User-provided name of output file
-    #[serde(rename = "name", skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    /// Optional users provided description field, can be used for documentation.
-    #[serde(rename = "description", skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    /// URL at which the TES server makes the output accessible after the task is complete. When tesOutput.path contains wildcards, it must be a directory; see `tesOutput.path_prefix` for details on how output URLs are constructed in this case. For Example:  - `s3://my-object-store/file1`  - `gs://my-bucket/file2`  - `file:///path/to/my/file`
-    #[serde(rename = "url")]
-    pub url: String,
-    /// Absolute path of the file inside the container. May contain pattern matching wildcards to select multiple outputs at once, but mind implications for `tesOutput.url` and `tesOutput.path_prefix`. Only wildcards defined in IEEE Std 1003.1-2017 (POSIX), 12.3 are supported; see https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_13
-    #[serde(rename = "path")]
-    pub path: String,
-    /// Prefix to be removed from matching outputs if `tesOutput.path` contains wildcards; output URLs are constructed by appending pruned paths to the directory specfied in `tesOutput.url`. Required if `tesOutput.path` contains wildcards, ignored otherwise.
-    #[serde(rename = "path_prefix", skip_serializing_if = "Option::is_none")]
-    pub path_prefix: Option<String>,
-    #[serde(rename = "type", skip_serializing_if = "Option::is_none")]
-    pub r#type: Option<models::TesFileType>,
+pub struct TesServiceType {
+    /// Namespace in reverse domain name format. Use `org.ga4gh` for implementations compliant with official GA4GH specifications. For services with custom APIs not standardized by GA4GH, or implementations diverging from official GA4GH specifications, use a different namespace (e.g. your organization's reverse domain name).
+    #[serde(rename = "group")]
+    pub group: String,
+    #[serde(rename = "artifact")]
+    pub artifact: Artifact,
+    /// Version of the API or specification. GA4GH specifications use semantic versioning.
+    #[serde(rename = "version")]
+    pub version: String,
 }
 
-impl TesOutput {
-    /// Output describes Task output files.
-    pub fn new(url: String, path: String) -> TesOutput {
-        TesOutput {
-            name: None,
-            description: None,
-            url,
-            path,
-            path_prefix: None,
-            r#type: None,
+impl TesServiceType {
+    pub fn new(group: String, artifact: Artifact, version: String) -> TesServiceType {
+        TesServiceType {
+            group,
+            artifact,
+            version,
         }
+    }
+}
+/// 
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
+pub enum Artifact {
+    #[serde(rename = "tes")]
+    Tes,
+}
+
+impl Default for Artifact {
+    fn default() -> Artifact {
+        Self::Tes
     }
 }
